Fixed phpcs errors manually
