+++ layers/GraphQLByPoP/packages/graphql-parser/src/Parser/Ast/ArgumentValue/Variable.php
@@ -49,7 +49,7 @@ class Variable extends AbstractAst implements ValueInterface
      * @param bool     $arrayElementNullable
      * @param Location $location
      */
-    public function __construct($name, $type, $nullable, $isArray, $arrayElementNullable = true, Location $location)
+    public function __construct($name, $type, $nullable, $isArray, $arrayElementNullable, Location $location)
     {
         parent::__construct($location);
