+++ layers/GraphQLByPoP/packages/graphql-parser/src/Parser/Ast/Query.php
@@ -36,7 +36,7 @@ class Query extends AbstractAst implements FieldInterface
      * @param array    $directives
      * @param Location $location
      */
-    public function __construct($name, $alias = '', array $arguments, array $fields, array $directives, Location $location)
+    public function __construct($name, $alias, array $arguments, array $fields, array $directives, Location $location)
     {
         parent::__construct($location);
