+++ layers/Schema/packages/everythingelse/src/FieldResolvers/CatEventFieldResolver.php
@@ -13,7 +13,7 @@ use PoPSchema\Events\TypeResolvers\EventTypeResolver;
 use PoP\ComponentModel\Misc\GeneralUtils;
 use PoPSchema\Events\Facades\EventTypeAPIFacade;
 
-class Cat_EventFieldResolver extends AbstractDBDataFieldResolver
+class CatEventFieldResolver extends AbstractDBDataFieldResolver
 {
     public static function getClassesToAttachTo(): array
     {
